+++ src/DashboardUI/package.json
@@ -31,7 +31,6 @@
     "react-compound-slider": "^3.3.1",
     "react-dnd": "^15.1.1",
     "react-dnd-touch-backend": "^15.1.1",
-    "react-dom": "^17.0.2",
     "react-multiselect-dropdown-bootstrap": "^1.1.0",
     "react-query": "^3.34.15",
     "react-router-dom": "^6.2.1",
