+++ src/DashboardUI/src/config/constants.ts
@@ -12,3 +12,18 @@ export const nldi = {
   },
   latLongPrecision: 4
 }
+
+export enum waterRightsProperties {
+  owners = "o",
+  ownerClassifications = "oClass",
+  beneficialUses = "bu",
+  siteUuid = "uuid",
+  sitePodOrPou = "podPou",
+  waterSourceTypes = "wsType",
+  minFlowRate = "minFlow",
+  maxFlowRate = "maxFlow",
+  minVolume = "minVol",
+  maxVolume = "maxVol",
+  minPriorityDate = "minPri",
+  maxPriorityDate = "maxPri",
+}
